refactor(frontend): extract form to a separate package
