+++ frontend/packages/form/package.json
@@ -1,0 +1,76 @@
+{
+  "name": "@vaadin/form",
+  "version": "22.0.0-alpha1",
+  "description": "Vaadin form utils",
+  "main": "./index.js",
+  "module": "./index.js",
+  "repository": "vaadin/flow/client",
+  "keywords": [
+    "Vaadin",
+    "Fusion"
+  ],
+  "scripts": {
+    "build": "tsc --isolatedModules",
+    "lint": "eslint --fix src/**/*.ts",
+    "test": "web-test-runner \"test/**/*.test.ts\" --config ../../web-test-runner.config.js",
+    "test:watch": "yarn run test -- --watch",
+    "typecheck": "tsc --noEmit"
+  },
+  "exports": {
+    ".": {
+      "default": "./index.js"
+    },
+    "./Binder.js": {
+      "default": "./Binder.js"
+    },
+    "./BinderNode.js": {
+      "default": "./BinderNode.js"
+    },
+    "./Field.js": {
+      "default": "./Field.js"
+    },
+    "./index.js": {
+      "default": "./index.js"
+    },
+    "./Models.js": {
+      "default": "./Models.js"
+    },
+    "./Validation.js": {
+      "default": "./Validation.js"
+    },
+    "./Validators.js": {
+      "default": "./Validators.js"
+    }
+  },
+  "author": "Vaadin Ltd",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/vaadin/fusion/issues"
+  },
+  "homepage": "https://vaadin.com/fusion",
+  "files": [
+    "Binder.{d.ts.map,d.ts,js.map,js}",
+    "BinderNode.{d.ts.map,d.ts,js.map,js}",
+    "Field.{d.ts.map,d.ts,js.map,js}",
+    "index.{d.ts.map,d.ts,js.map,js}",
+    "Models.{d.ts.map,d.ts,js.map,js}",
+    "Validation.{d.ts.map,d.ts,js.map,js}",
+    "Validators.{d.ts.map,d.ts,js.map,js}"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "dependencies": {
+    "lit": "^2.0.0-rc.2",
+    "validator": "^13.6.0"
+  },
+  "devDependencies": {
+    "@open-wc/testing": "^2.5.33",
+    "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
+    "@types/sinon": "^10.0.2",
+    "@types/sinon-chai": "^3.2.5",
+    "@web/test-runner": "^0.13.16",
+    "sinon": "^11.1.2"
+  }
+}
