+++ frontend/packages/form/test/Model.test.ts
@@ -12,7 +12,7 @@ import {
   NumberModel,
   Positive,
   Size,
-} from '../../src';
+} from '../src';
 
 import { IdEntity, IdEntityModel, TestEntity, TestModel } from './TestModels';
