+++ frontend/packages/form/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "."
+  },
+  "include": ["src"],
+  "exclude": ["!src/**/*"]
+}
